Add some C++ protection to the libcap/execable.* code.
These ideas were what I shared in this answer:

  https://stackoverflow.com/a/68339111/14760867

so I figure I might as well capture them here. Credit goes to the stackoverflow user, "Haydentech", for running
into the issue and prompting me to figure out the issue.

All that being said, within the libcap build sources, is not completely
clear when the __cplusplus support would be invoked.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>