libcap: Improve raw container calculation
On architectures with pointer alignment constraints larger than 8
bytes, e.g. CHERI RISC-V, a raw container calculation in libcap
results in invalid read accesses to members or memory corruption.
Fix this by taking the alignment constraints into account as well
as asserting the memory layout of the container struct.

For capability allocation, libcap uses the struct:
	struct _cap_alloc_s {
	    __u32 magic;
	    __u32 size;
	    union {
		...
    	    } u;
	};

where the union abstracts several capability types that can be
identified by _cap_alloc_s' magic member. To obtain a pointer to the
allocation struct from a pointer to the union, the following
expression is used:
	struct _cap_alloc_s *p = (void *) (-2 + (__u32 *) union_ptr);

This calculation relies on the assumption that the union is placed at an
offset of 8 bytes in the allocation struct. However, some struct types
that are in the union may have pionters and thus, these structs may have
larger alignment constraints on architectures with fat pointers, e.g.
16-byte alignment on CHERI RISC-V. In these cases, the offset may be
larger than 8 bytes and the raw container calculation results in an
invalid pointer that is dereferenced.

Fix the caluculation by using the alignment constraint of the union if
it exceeds the size of the two preceeding 32-bit uints magic and size.
Statically assert that the so calculated offset matches the actual
offset of u in _cap_alloc_s.

Signed-off-by: Chris Hofer <christian.hofer@codasip.com>

Added compilation fix for -std=c99 compilation, and suppressed
newly introduced warnings from clang.

Resolves:

  https://bugzilla.kernel.org/show_bug.cgi?id=220415

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>